+++ src/main.rs
@@ -2,19 +2,22 @@ use std::env;
 
 mod api;
 mod script;
-mod update;
+mod helper;
 mod utils;
 
 #[cfg(windows)]
 mod windows_utf8;
 
-pub const VERSION: &str = "0.1.8";
+pub const VERSION: &str = "0.1.9";
 
-pub fn deprecated(name: &str, version: &str) {
-    println!("[DEPRECATED-WARNING] '{}' is deprectaed since Version: {}", name, version);
-}
+use crate::helper::print::{
+    END,
+    RED,
+    GREEN
+};
 
 fn main() {
+    // Must be add the start IMPORTANT
     #[cfg(windows)]
     let _ = windows_utf8::enable_utf8();
 
@@ -22,11 +25,12 @@ fn main() {
 
     // Check if filename
     if args.len() < 2 {
-        eprintln!("No Path provided! Run with -h or --help for more information.");
+        eprintln!("{}No Path provided! Run with -h or --help for more information.{}", RED, END);
         std::thread::sleep(std::time::Duration::from_secs(5));
         return;
     }
 
+    // Help
     if args[1] == "-h" || args[1] == "--help" {
         println!("LuaAPI-Rust {}", VERSION);
         println!("Usage: luaapi-rust <script.lua> [-safe]");
@@ -41,20 +45,20 @@ fn main() {
         // Save Mode
         if args[2] == "-safe" {
             safe = true;
-            eprintln! {"Safe is not implemnted yet"}
+            eprintln!("{}[ERROR] Safe is not implemnted yet{}", RED, END);
             std::thread::sleep(std::time::Duration::from_secs(5));
             return;
         } else {
             if args[1] == "run" {
                 // Update
                 if args[2] == "update" {
-                    if let Err(e) = update::update() {
+                    if let Err(e) = helper::update::update() {
                         eprintln!("Update error: {}", e);
                     }
                 }
                 // Install
                 else if args[2] == "install" {
-                    if let Err(e) = update::install() {
+                    if let Err(e) = helper::update::install() {
                         eprintln!("Installation error: {}", e);
                     }
                 }
@@ -67,11 +71,15 @@ fn main() {
 
     let path = &args[1];
 
-    println!("[LUAJIT-INFO] running script: {}", path);
+    println!("{}[LUAJIT-INFO] running script: {}{}", GREEN, path, END);
 
     if let Err(e) = script::execute_script(path, &safe) {
-        eprintln!("Script error: {}", e);
+        eprintln!("{}[LUAJIT-ERROR] Script error: {}{}", RED, e, END);
         std::thread::sleep(std::time::Duration::from_secs(5));
         return;
     }
+
+    println!("{}[LUAJIT-INFO] finished script executing: {}{}", GREEN, path, END);
+
+    helper::print::print();
 }